backend-app/requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
pydantic
python-dotenv
python-jose
email-validator